Stable formatting between Jackson 2.16 and earlier versions
Unfortunatily this reaches even deeper into Jackson internals than
before, so it might break in the future (again). On the other hand,
this sidesteps messing with reflection or other hacks to decide on which
version of Jackson we are currently running.
