+++ java-snapshot-testing-plugin-jackson/build.gradle
@@ -21,10 +21,10 @@ dependencies {
     testImplementation 'org.assertj:assertj-core:3.11.1'
     testImplementation 'org.skyscreamer:jsonassert:1.5.0' // For docs/ reporter example
 
-    testImplementation 'com.fasterxml.jackson.core:jackson-core:2.11.3'
-    testImplementation 'com.fasterxml.jackson.core:jackson-databind:2.11.3'
-    testRuntimeOnly 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.11.3'
-    testRuntimeOnly 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.11.3'
+    testImplementation 'com.fasterxml.jackson.core:jackson-core:2.16.0'
+    testImplementation 'com.fasterxml.jackson.core:jackson-databind:2.16.0'
+    testRuntimeOnly 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.16.0'
+    testRuntimeOnly 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.16.0'
 }
 
 test { useJUnitPlatform() }
